refactor: typo package
