+++ lib/reader/common/bytes.go
