+++ lib/reader/common/match/optimize.go
