+++ lib/reader/common/streambuf/ascii.go
