+++ lib/reader/multiline/while.go
@@ -20,7 +20,7 @@ package multiline
 import (
 	logger "github.com/cihub/seelog"
 	"infini.sh/agent/lib/reader"
-	"infini.sh/agent/lib/reader/commom/match"
+	"infini.sh/agent/lib/reader/common/match"
 	"infini.sh/agent/lib/reader/readfile"
 	"io"
 )
